+++ Changelog.md
@@ -1,6 +1,8 @@
 # NeuroScheme
 
 ## git master
+* [!11] Added nsol Xml scene loading via CLI.
+* [!9] Added CLI help message.
 * [!8] Enabled building in Windows without console. Separate lib, GUI app and plugins.
 * [!7] Bumped dependencies versions. Updated README.md
 * [!6] Added home button. Added circular layout. Fixed bugs.